Implement splash screen functionality with SOLID principles, including navigation, initialization, and timer services. Replace old splash screen implementation with new structured approach.
